+++ CHANGELOG.markdown
@@ -3,6 +3,9 @@
 
 ## Version 0.9.2 - *Not Yet Released*
 
+ * [Tools] Added an "aphid" command-line tool that can be used to bootstrap a
+   new Aphid application (e.g. "aphid new ProjectName").
+
  * [Skeleton] Added a skeleton project in the Skeleton directory that may be
    used to start new Aphid-based applications. It contains an application
    delegate, a view controller and a Rakefile for building the project.